+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- A typo in the `setup.cfg` file (#24)
+
 ## [1.2.0] - 2019-08-12
 
 ### Added